pyproject.toml
[tool.poetry]
name = "multiverse"
version = "0.1.0"
description = "Multiverse Tech Task - Mars Rover"
authors = ["Miguel dos Santos Veríssimo Noronha <msvn91@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest-mock = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
